Move creating readme into own script
